Left View of Binary Tree
Given a Binary Tree, print the Left view of it. The left view of a Binary Tree is a set of nodes visible when the tree is visited from the Left side. The task is to complete the function leftView(), which accepts the root of the tree as an argument.