Next Greater Element
Given an array arr[ ] of size N having distinct elements, the task is to find the next greater element for each element of the array in order of their appearance in the array.
The next greater element of an element in the array is the nearest element on the right which is greater than the current element.
If there does not exist the next greater of the current element, then the next greater element for the current element is -1. For example, the next greater of the last element is always -1.